Disambiguate package names
This is beneficial for the JaCoCo report that is now generated on root level, and would not distinguish subprojects by their name.

Also rename `com.vary` to `com.adnovum`, to reflect the recent project move.
