+++ base/src/main/java/com/adnovum/tenantsecrets/base/crypto/CryptoException.java
@@ -1,0 +1,12 @@
+package com.adnovum.tenantsecrets.base.crypto;
+
+public class CryptoException extends Exception {
+
+	public CryptoException(String message) {
+		super(message);
+	}
+
+	public CryptoException(String message, Throwable cause) {
+		super(message, cause);
+	}
+}
