+++ base/src/main/java/com/adnovum/tenantsecrets/base/crypto/Encrypter.java
@@ -1,0 +1,7 @@
+package com.adnovum.tenantsecrets.base.crypto;
+
+public interface Encrypter {
+
+	String encrypt(byte[] key, String plainText) throws CryptoException;
+	String decrypt(byte[] key, String cipherText) throws CryptoException;
+}
