+++ base/src/main/java/com/adnovum/tenantsecrets/base/crypto/KeyDeriver.java
@@ -1,0 +1,6 @@
+package com.adnovum.tenantsecrets.base.crypto;
+
+public interface KeyDeriver {
+
+	byte[] deriveKey(byte[] originalKey, String context);
+}
