--- base/src/main/java/com/vary/tenantsecrets/crypto/CryptoException.java
@@ -1,12 +1,0 @@
-package com.vary.tenantsecrets.crypto;
-
-public class CryptoException extends Exception {
-
-	public CryptoException(String message) {
-		super(message);
-	}
-
-	public CryptoException(String message, Throwable cause) {
-		super(message, cause);
-	}
-}
