--- base/src/main/java/com/vary/tenantsecrets/crypto/KeyDeriver.java
@@ -1,6 +1,0 @@
-package com.vary.tenantsecrets.crypto;
-
-public interface KeyDeriver {
-
-	byte[] deriveKey(byte[] originalKey, String context);
-}
