+++ plugin/src/main/resources/secrets.template.html
@@ -14,7 +14,7 @@
   ~ limitations under the License.
   -->
 <!-- Taken and adapted from github.com/gocd-contrib/secrets-skeleton-plugin. -->
-<div data-plugin-style-id="com.vary.tenantsecrets">
+<div data-plugin-style-id="com.adnovum.tenantsecrets.plugin">
 
     <div class="form_item_block">
         <label>Tenant identifier:<span class="asterisk">*</span></label>
