+++ webservice/src/main/java/com/adnovum/tenantsecrets/webservice/TenantsecretsApplication.java
@@ -1,0 +1,13 @@
+package com.adnovum.tenantsecrets.webservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TenantsecretsApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TenantsecretsApplication.class, args);
+	}
+
+}
