--- webservice/src/main/java/com/vary/tenantsecrets/tenantsecrets/TenantsecretsApplication.java
@@ -1,13 +1,0 @@
-package com.vary.tenantsecrets.tenantsecrets;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TenantsecretsApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TenantsecretsApplication.class, args);
-	}
-
-}
